Modify sequence of pages for match to collaboration
